All this does is cleanup a bunch of style issues that Resharper complains about (#629)
* Removed unused using statements

* Removed redudnant ToString()'s

* Removed some redundant type casts

* Removed redundant bool comparisons

* remove redundant initializers

* xml summary in invalid place switched to comments

* Use string interpolation

* use format separators

* Removed redundant parenths

* remove redundant else

* direct cast when safe

* redundant string interpolation

* xml summary fixes

* removed empty ctors

* fix modifier order

* Convert to auto property

* Convert some getters to method bodies

* use format separators

* changelog

* Auto-properties that can made to get-only

* fixed some references in XML comments

* remove redundant base()

* Revert "Auto-properties that can made to get-only"

This reverts commit 32a1225ce8a9e185aae38067ba2bd521488267ec.

* fix starter gear json

* Use collections count property

* simplify conditional ternary expression

* remove redundant return statements

* Constructors for abstract classes changed to protected

* Moved declarations to inner scopes

* Join declaration and assignments

* inline out variables

* Use enum extension methods instead of static methods

* inline out more variables

* more unused usings

* Remove specifying enums as ints

+++ Source/ACE.Database/WorldDatabase.cs
@@ -1,12 +1,10 @@
 using System;
-using System.Threading.Tasks;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 using MySql.Data.MySqlClient;
 using ACE.Entity;
 using ACE.Entity.Enum;
-using System.Diagnostics;
 
 namespace ACE.Database
 {
@@ -270,7 +268,7 @@ namespace ACE.Database
                 where = where != null ? where + " AND " : "";
                 where += "`v`.aceObjectId = ?";
                 var p = new MySqlParameter("", MySqlDbType.UInt32);
-                p.Value = (uint)criteria.WeenieClassId.Value;
+                p.Value = criteria.WeenieClassId.Value;
                 mysqlParams.Add(p);
             }
